Working lowpass
Still some frame drop. I think it's due to the buffer size being too small
while recalculating the cutoff frequency rapidly.

I expect increasing the buffer size will reduce the audible drop from hardware
lag, but at the expense of response to the knob's position with respect to
time. I will attempt to tune a sweetspot if possible where audio drop is
eliminated and response lag is minimal.

I also would like to improve the UI with proper labels, more controls, and a
filter depth, on/off switch, and resonance control. These improvements will
probably be late, but I would just like to submit this commit this as a
timestamp of completion.